+++ Product.Model/ProductOfferModel.cs
@@ -17,7 +17,7 @@ namespace Product.Model
         public string ProductName { get; set; }
         public Nullable<bool> Status { get; set; }
         public string Message { get; set; }
-        public Nullable<decimal> Counter { get; set; }
+        public Nullable<int> Counter { get; set; }
         public virtual IEnumerable<UserModel> UserModels { get; set; }
         public virtual IEnumerable<ProductModel> ProductModels { get; set; }
     }
